apps/pipelines/requirements.txt
aiohttp>=3.9,<4
supabase>=2.0,<3
python-dotenv>=1.0,<2
tenacity>=8.2,<9
pydantic>=2.0,<3
